% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfun.R
\name{edfun}
\alias{edfun}
\title{Title}
\usage{
edfun(x, ...)
}
\arguments{
\item{x}{numeric vector of data}

\item{...}{}
}
\value{
A list with 4 components: dfun, pfun, qfun and rfun.
Each one is a function to perform the usual tasks of distributions.
}
\description{
Title
}
\examples{

set.seed(2016-08-18)
x <- rnorm(100)
x_funs <- edfun(x)

f <- x_funs$dfun
curve(f, -2,2)

f <- x_funs$pfun
curve(f, -2,2)

f <- x_funs$qfun
curve(f, 0,1)

f <- x_funs$rfun
hist(f(1000))


}

