% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfun.R
\name{edfun}
\alias{edfun}
\title{Creating Empirical Distribution Functions}
\usage{
edfun(x, support = range(x), ...)
}
\arguments{
\item{x}{numeric vector of data}

\item{support}{a 2d numeric vector giving the boundaries of the distribution.
Default is the range of x.
This is used in qfun to decide how to work with extreme cases of q->0|1.}

\item{...}{}
}
\value{
A list with 4 components: dfun, pfun, qfun and rfun.
Each one is a function to perform the usual tasks of distributions.
}
\description{
Creating Empirical Distribution Functions
}
\examples{

set.seed(2016-08-18)
x <- rnorm(100)
x_funs <- edfun(x)
x_funs$qfun(0) # -2.6

# for extreme cases, we can add the support vector
x_funs <- edfun(x, support = c(-Inf, Inf))
x_funs$qfun(0) # -Inf

f <- x_funs$dfun
curve(f, -2,2)

f <- x_funs$pfun
curve(f, -2,2)

f <- x_funs$qfun
curve(f, 0,1)

f <- x_funs$rfun
hist(f(1000))

}

